Bilan de la journée (20/04/2026) Aujourd'hui a été une journée de "débroussaillage" et de standardisation. Succès de la compilation : Le main.c, lcd.c, graphics.c et serial.c compilent désormais sans erreur sur l'architecture ARM Cortex-M0+ du STM32. Standardisation des types : Éradication des conflits de types (u_int obsolète) au profit des standards robustes du C (uint16_t, uint8_t), ce qui garantit la bonne gestion des couleurs 16 bits. Nettoyage des avertissements : Application d'un #pragma pour forcer le compilateur à accepter la structure du fichier fonts.h sans polluer la console. Liaison des bibliothèques : Le main.c reconnait enfin les fonctions de l'écran grâce aux bonnes inclusions (#include "lcd.h", etc.). Tâche du Cahier des ChargesProgressionAnalyse objective1. Affichage des écrans TFT95 %Le code bas niveau compile. L'étape manquante est la validation physique/visuelle (voir le rendu réel).2. Insertion des données (TFT)100 %✅ graphics.c et les polices sont 100% compatibles STM32. Prêt à dessiner.3. Passage d'un écran via série60 %Le HAL UART est en place, mais la logique applicative de réception/interprétation des données reste à tester en situation réelle.4. Mise en veille (10s)85 %Schéma décodé : La broche ALIM_AFF pilote la base d'un transistor PNP (BC856A). L'écran s'allume quand on met la broche à 0 (LOW) et s'éteint à 1 (HIGH).5. Touches capacitives réelles0 %En attente.6. Éphéméride (Heure + GPS)0 %En attente de tes anciens fichiers.

+++ Tarek_Remote_Screen/Core/Inc/graphics.h
@@ -21,8 +21,8 @@
 #define FONT_MD_BKG	4
 #define FONT_LG_BKG	5
 //--------Ancien-------
-/*void setColor(u_int color);
-void setBackgroundColor(u_int color);*/
+/*void setColor(uint16_t color);
+void setBackgroundColor(uint16_t color);*/
 //--------Nouveau------
 void setColor(uint16_t color);
 void setBackgroundColor(uint16_t color);
